+++ aip-console-jenkins/src/main/java/io/jenkins/plugins/aipconsole/AddVersionBuilder.java
@@ -19,6 +19,7 @@ import com.castsoftware.aip.console.tools.core.services.JobsService;
 import com.castsoftware.aip.console.tools.core.services.RestApiService;
 import com.castsoftware.aip.console.tools.core.services.UploadService;
 import com.castsoftware.aip.console.tools.core.utils.Constants;
+import com.castsoftware.aip.console.tools.core.utils.VersionObjective;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
@@ -113,6 +114,7 @@ public class AddVersionBuilder extends BaseActionBuilder implements SimpleBuildS
 
     @Nullable
     private String snapshotName = "";
+    private boolean blueprint = false;
 
     @DataBoundConstructor
     public AddVersionBuilder(String applicationName, String filePath) {
@@ -152,6 +154,19 @@ public class AddVersionBuilder extends BaseActionBuilder implements SimpleBuildS
     @DataBoundSetter
     public void setAutoCreate(boolean autoCreate) {
         this.autoCreate = autoCreate;
+    }
+
+    @DataBoundSetter
+    public void setBlueprint(boolean blueprint) {
+        this.blueprint = blueprint;
+    }
+
+    public boolean getBlueprint() {
+        return isBlueprint();
+    }
+
+    public boolean isBlueprint() {
+        return blueprint;
     }
 
     public boolean isCloneVersion() {
@@ -267,7 +282,7 @@ public class AddVersionBuilder extends BaseActionBuilder implements SimpleBuildS
     public void perform(@Nonnull Run<?, ?> run, @Nonnull FilePath workspace, @Nonnull Launcher launcher, @Nonnull TaskListener listener) throws InterruptedException, IOException {
         PrintStream log = listener.getLogger();
         Result defaultResult = failureIgnored ? Result.UNSTABLE : Result.FAILURE;
-        boolean applicationHasVersion = this.cloneVersion;
+        boolean applicationHasVersion = cloneVersion;
         boolean isUpload = false;
 
         String errorMessage;
@@ -357,19 +372,20 @@ public class AddVersionBuilder extends BaseActionBuilder implements SimpleBuildS
                     return;
                 }
                 // Is there a node name
+                String expandedNodeName = vars.expand(nodeName);
                 String nodeGuid = null;
-                if (StringUtils.isNotBlank(nodeName)) {
+                if (StringUtils.isNotBlank(expandedNodeName)) {
                     try {
                         nodeGuid = apiService.getForEntity("/api/nodes",
                                 new TypeReference<List<NodeDto>>() {
                                 }).stream()
-                                .filter(n -> StringUtils.equalsIgnoreCase(n.getName(), nodeName))
+                                .filter(n -> StringUtils.equalsIgnoreCase(n.getName(), expandedNodeName))
                                 .map(NodeDto::getGuid)
                                 .findFirst()
                                 .orElse(null);
 
                         if (StringUtils.isBlank(nodeGuid)) {
-                            listener.error(AddVersionBuilder_AddVersion_error_nodeNotFound(nodeName));
+                            listener.error(AddVersionBuilder_AddVersion_error_nodeNotFound(expandedNodeName));
                             run.setResult(defaultResult);
                             return;
                         }
@@ -470,7 +486,7 @@ public class AddVersionBuilder extends BaseActionBuilder implements SimpleBuildS
                 resolvedVersionName = String.format("v%s", formatVersionName.format(new Date()));
             }
 
-            if (this.cloneVersion) {
+            if (cloneVersion) {
                 if (applicationHasVersion) {
                     log.println(AddVersionBuilder_AddVersion_info_startCloneVersionJob(variableAppName));
                 } else {
@@ -492,6 +508,9 @@ public class AddVersionBuilder extends BaseActionBuilder implements SimpleBuildS
                 requestBuilder.deliveryConfigGuid(deliveryConfig);
             }
 
+            if (isBlueprint()) {
+                requestBuilder.objectives(VersionObjective.BLUEPRINT);
+            }
             if (StringUtils.isNotBlank(resolvedSnapshotName)) {
                 requestBuilder.snapshotName(resolvedSnapshotName);
             }