+++ aip-console-jenkins/src/main/java/io/jenkins/plugins/aipconsole/DeliverBuilder.java
@@ -20,6 +20,7 @@ import com.castsoftware.aip.console.tools.core.services.JobsService;
 import com.castsoftware.aip.console.tools.core.services.RestApiService;
 import com.castsoftware.aip.console.tools.core.services.UploadService;
 import com.castsoftware.aip.console.tools.core.utils.Constants;
+import com.castsoftware.aip.console.tools.core.utils.VersionObjective;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
@@ -119,6 +120,7 @@ public class DeliverBuilder extends BaseActionBuilder implements SimpleBuildStep
 
     private boolean autoDiscover = true;
     private boolean setAsCurrent = false;
+    private boolean blueprint = false;
 
     @DataBoundConstructor
     public DeliverBuilder(String applicationName, String filePath) {
@@ -188,6 +190,19 @@ public class DeliverBuilder extends BaseActionBuilder implements SimpleBuildStep
 
     public boolean getSetAsCurrent() {
         return isSetAsCurrent();
+    }
+
+    @DataBoundSetter
+    public void setBlueprint(boolean blueprint) {
+        this.blueprint = blueprint;
+    }
+
+    public boolean getBlueprint() {
+        return isBlueprint();
+    }
+
+    public boolean isBlueprint() {
+        return blueprint;
     }
 
     @DataBoundSetter
@@ -291,7 +306,7 @@ public class DeliverBuilder extends BaseActionBuilder implements SimpleBuildStep
     public void perform(@Nonnull Run<?, ?> run, @Nonnull FilePath workspace, @Nonnull Launcher launcher, @Nonnull TaskListener listener) throws InterruptedException, IOException {
         PrintStream log = listener.getLogger();
         Result defaultResult = failureIgnored ? Result.UNSTABLE : Result.FAILURE;
-        boolean applicationHasVersion = this.cloneVersion;
+        boolean applicationHasVersion = cloneVersion;
         boolean isUpload = false;
 
         String errorMessage;
@@ -383,18 +398,19 @@ public class DeliverBuilder extends BaseActionBuilder implements SimpleBuildStep
                 }
                 // Is there a node name
                 String nodeGuid = null;
-                if (StringUtils.isNotBlank(nodeName)) {
+                String expandedNodeName = vars.expand(nodeName);
+                if (StringUtils.isNotBlank(expandedNodeName)) {
                     try {
                         nodeGuid = apiService.getForEntity("/api/nodes",
                                 new TypeReference<List<NodeDto>>() {
                                 }).stream()
-                                .filter(n -> StringUtils.equalsIgnoreCase(n.getName(), nodeName))
+                                .filter(n -> StringUtils.equalsIgnoreCase(n.getName(), expandedNodeName))
                                 .map(NodeDto::getGuid)
                                 .findFirst()
                                 .orElse(null);
 
                         if (StringUtils.isBlank(nodeGuid)) {
-                            listener.error(AddVersionBuilder_AddVersion_error_nodeNotFound(nodeName));
+                            listener.error(AddVersionBuilder_AddVersion_error_nodeNotFound(expandedNodeName));
                             run.setResult(defaultResult);
                             return;
                         }
@@ -496,7 +512,7 @@ public class DeliverBuilder extends BaseActionBuilder implements SimpleBuildStep
                 resolvedVersionName = String.format("v%s", formatVersionName.format(new Date()));
             }
 
-            if (this.cloneVersion) {
+            if (cloneVersion) {
                 if (applicationHasVersion) {
                     log.println(DeliverBuilder_Deliver_info_startDeliverCloneJob(expandedAppName));
                 } else {
@@ -516,6 +532,10 @@ public class DeliverBuilder extends BaseActionBuilder implements SimpleBuildStep
 
             if (inplaceMode || isSetAsCurrent()) {
                 requestBuilder.endStep(Constants.SET_CURRENT_STEP_NAME);
+            }
+            
+            if (isBlueprint()) {
+                requestBuilder.objectives(VersionObjective.BLUEPRINT);
             }
 
             log.println("Exclusion patterns : " + exclusionPatterns);