+++ aip-console-tools-cli/src/main/java/com/castsoftware/aip/console/tools/commands/AnalyzeCommand.java
@@ -13,6 +13,8 @@ import com.castsoftware.aip.console.tools.core.exceptions.ApiKeyMissingException
 import com.castsoftware.aip.console.tools.core.exceptions.ApplicationServiceException;
 import com.castsoftware.aip.console.tools.core.exceptions.JobServiceException;
 import com.castsoftware.aip.console.tools.core.services.ApplicationService;
+import com.castsoftware.aip.console.tools.core.services.DebugOptionsService;
+import com.castsoftware.aip.console.tools.core.services.DebugOptionsServiceImpl;
 import com.castsoftware.aip.console.tools.core.services.JobsService;
 import com.castsoftware.aip.console.tools.core.services.RestApiService;
 import com.castsoftware.aip.console.tools.core.utils.ApiEndpointHelper;
@@ -21,6 +23,7 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import picocli.CommandLine;
 
@@ -87,6 +90,9 @@ public class AnalyzeCommand implements Callable<Integer> {
             , defaultValue = "false")
     private boolean amtProfiling;
 
+    @Autowired
+    private DebugOptionsService debugOptionsService;
+
     public AnalyzeCommand(RestApiService restApiService, JobsService jobsService, ApplicationService applicationService) {
         this.restApiService = restApiService;
         this.jobsService = jobsService;
@@ -112,6 +118,7 @@ public class AnalyzeCommand implements Callable<Integer> {
         }
         String applicationGuid;
         ApiInfoDto apiInfoDto = restApiService.getAipConsoleApiInfo();
+
         log.info("[Debug options] Show Sql is '{}'", showSql);
         log.info("[Debug options] AMT Profiling is '{}'", amtProfiling);
 
@@ -163,16 +170,8 @@ public class AnalyzeCommand implements Callable<Integer> {
                     .versionGuid(versionToAnalyze.getGuid())
                     .releaseAndSnapshotDate(new Date());
 
-            //==============
-            // Ony set debug option when ON. Run Analysis always consider these options as OFF(disabled)
-            //==============
-            DebugOptionsDto oldDebugOptions = applicationService.getDebugOptions(applicationGuid);
-            if (showSql) {
-                applicationService.updateShowSqlDebugOption(applicationGuid, showSql);
-            }
-            if (amtProfiling) {
-                applicationService.updateShowSqlDebugOption(applicationGuid, amtProfiling);
-            }
+            DebugOptionsDto oldDebugOptions = debugOptionsService.updateDebugOptions(applicationGuid,
+                    DebugOptionsDto.builder().showSql(showSql).activateAmtMemoryProfile(amtProfiling).build());
 
             log.info("Running analysis for application '{}' with version '{}'", applicationName, versionToAnalyze.getName());
             String jobGuid = jobsService.startJob(builder);
@@ -182,8 +181,8 @@ public class AnalyzeCommand implements Callable<Integer> {
             JobStatusWithSteps jobStatus = jobsService.pollAndWaitForJobFinished(jobGuid, Function.identity(), sharedOptions.isVerbose());
             Runtime.getRuntime().removeShutdownHook(shutdownHook);
 
-            DebugOptionsDto debugOptions = applicationService.getDebugOptions(applicationGuid);
-            applicationService.resetDebugOptions(applicationGuid, oldDebugOptions);
+            DebugOptionsDto debugOptions = debugOptionsService.getDebugOptions(applicationGuid);
+            debugOptionsService.resetDebugOptions(applicationGuid, oldDebugOptions);
             if (JobState.COMPLETED == jobStatus.getState()) {
                 if (debugOptions.isActivateAmtMemoryProfile()) {
                     log.info("[Debug options] Amt Profiling file download URL: {}",